refactored code
rewrote pointer arithmetic regarding function calls.

move common code between AES.c and AES_smallScaleVersion.c into
own AES_common.c file and adjust AES.h (now AES_common.h).
Work on Issue #10 and partly on Issue #9 (function names are translated)
